Replace GitHub PAT authentication with GitHub App
Personal Access Tokens are bound to a GitHub user, which doesn't work
for teams. And when the workflow fails, that GitHub user gets notified.

Using a GitHub App to authenticate makes it easier for teams to manage
access.

The unit test has been upgraded to a full integration test, that uses
a GitHub App to test. Faking the new authentication process makes the
unit test useless. Having a full integration tests hopefully helps, and
won't have to much false positives.